Add support for CC=gcc
Adds and overridable CC variable that defaults to the currently
hard-coded clang.

Also moves the linker flags to *after* the source file pattern match.
GCC will fail to compile due to undefined reference errors if the linker
flags come before a source file in which the associated functions are
used.